backend/requirements.txt
ollama
openai-whisper
gradio
pydub
soundfile
pyaudio
piper-tts
PyMuPDF
python-docx
reportlab
speechrecognition
duckduckgo-search
ddgs
matplotlib
regex
google-generativeai
flask
flask-cors